Implemented schema type resolution against the new TypeManager and Type classes
